+++ packages/three-elements/src/util/dasherize.ts
@@ -1,5 +1,5 @@
 export const dasherize = (str: string) =>
   str
-    .replace(/([a-z])([A-Z])/g, "$1-$2")
-    .replace(/([A-Z])([A-Z][a-z])/g, "$1-$2")
+    .replace(/([A-Za-z0-9])([A-Z][a-z])/g, "$1-$2")
+    .replace(/_/g, "-")
     .toLowerCase()