+++ packages/three-elements/test/custom-renderers.test.ts
@@ -23,28 +23,28 @@ describe("using custom renderers", () => {
     const render = () =>
       fixture(html`
         <three-game>
-          <three-web-gl-renderer xr.enabled></three-web-gl-renderer>
+          <three-webgl-renderer xr.enabled></three-webgl-renderer>
           <three-scene></three-scene>
         </three-game>
       `)
 
     it("creates an element that wraps a renderer", async () => {
       const game = (await render()) as ThreeGame
-      const renderer = game.querySelector("three-web-gl-renderer") as ThreeElement
+      const renderer = game.querySelector("three-webgl-renderer") as ThreeElement
 
       expect(renderer.object).to.be.instanceOf(THREE.WebGLRenderer)
     })
 
     it("automatically attaches the custom renderer to the three-game element", async () => {
       const game = (await render()) as ThreeGame
-      const renderer = game.querySelector("three-web-gl-renderer") as ThreeElement
+      const renderer = game.querySelector("three-webgl-renderer") as ThreeElement
 
       expect(game.renderer).to.eq(renderer.object)
     })
 
     it("passes custom options to the renderer", async () => {
       const game = (await render()) as ThreeGame
-      const renderer = game.querySelector("three-web-gl-renderer") as ThreeElement
+      const renderer = game.querySelector("three-webgl-renderer") as ThreeElement
 
       expect(renderer.object.xr.enabled).to.eq(true)
     })